Stop trying so hard; just use Set
Most of the time we are sampling our data, so it's OK to drop data if
memcache gets busy. Using CAS and retry logic means that it takes more
time for the caller to ensure that the data was correct, and that's time
we usually don't want to spend. So, just use Set and hope for the best.